train_model/requirements.txt
scipy.sparse
lightfm
pandas
numpy